clean up paged responses
