+++ rust/open-metadata/Cargo.toml
@@ -20,6 +20,7 @@ bytes = "1"
 async-trait = "0.1"
 rand = "*"
 reqwest-pipeline = { path = "../reqwest-pipeline" }
+futures-util = "0.3"
 
 [dependencies.reqwest]
 version = "^0.11"
