+++ rust/open-metadata/src/operations/collections.rs
@@ -3,7 +3,7 @@ use reqwest_pipeline::{collect_pinned_stream, Response};
 
 impl CollectionList {
     pub(crate) async fn try_from(response: Response) -> reqwest_pipeline::Result<Self> {
-        let (_status_code, headers, pinned_stream) = response.deconstruct();
+        let (_status_code, _headers, pinned_stream) = response.deconstruct();
         let body = collect_pinned_stream(pinned_stream).await?;
         Ok(serde_json::from_slice::<CollectionList>(&body)?)
     }
@@ -11,7 +11,7 @@ impl CollectionList {
 
 impl CatalogVersion {
     pub(crate) async fn try_from(response: Response) -> reqwest_pipeline::Result<Self> {
-        let (_status_code, headers, pinned_stream) = response.deconstruct();
+        let (_status_code, _headers, pinned_stream) = response.deconstruct();
         let body = collect_pinned_stream(pinned_stream).await?;
         Ok(serde_json::from_slice::<CatalogVersion>(&body)?)
     }
