+++ rust/open-metadata/src/operations/mod.rs
@@ -1,3 +1,5 @@
 mod collections;
+mod databases;
 
 pub use collections::*;
+pub use databases::*;
